fix: make equalTemperamentHz public; tighten implicit-conversion warnings
TuningClient: equalTemperamentHz moved from private to public so
TuningTests can call it directly.  Added explicit static_cast<float>
in the implementation to silence the int→float precision warning.

SynthVoice: cast totalPitchbendInSemitones to float before dividing
(it is double in JUCE) to silence the double→float precision warnings
introduced in the pitchbend-trill fix.

+++ Source/MPE/TuningClient.cpp
@@ -64,5 +64,5 @@ void TuningClient::reconnect()
 
 float TuningClient::equalTemperamentHz(int midiNote)
 {
-    return 440.0f * std::pow(2.0f, (midiNote - 69) / 12.0f);
+    return 440.0f * std::pow(2.0f, static_cast<float>(midiNote - 69) / 12.0f);
 }
